internal/broadcasts: fix test fluke
I'm not 100% sure how this ever passed so often. range over a map is
guaranteed to be random per the language spec. Not sure if this is
different for pointers but I'd be supprised.

Well, now this is testing the right things directly at least and doesn't
depend on the order of the output.